feat(election): added new election page
