+++ src/pages/dho/Home.vue
@@ -6,8 +6,6 @@ import I18n from '~/utils/i18n'
 
 import ipfsy from '~/utils/ipfsy'
 
-import formatNumber from '~/utils/formatNumber'
-
 const ordersMap = [{ asc: 'createdDate' }, { desc: 'createdDate' }, { asc: 'details_title_s' }]
 const UPVOTE_DOC_URL = 'https://help.hypha.earth/hc/2431449449/93/community-voting-method?category_id=42'
 
@@ -146,10 +144,7 @@ export default {
 
   data () {
     return {
-      formatNumber,
       I18n,
-      counterdown: undefined,
-      endDate: '2023-08-20',
       pagination: {
         first: 6,
         offset: 0,
@@ -158,11 +153,7 @@ export default {
       textFilter: null,
       order: ordersMap[0],
       currentUpvoteStep: null,
-      UPVOTE_DOC_URL,
-      timeRemaining: {},
-      participants: 126, // TODO: waiting API
-      treasury: 1200000,
-      isUpVoteElectionBannerVisible: true // TODO: waiting API
+      UPVOTE_DOC_URL
     }
   },
 
@@ -198,13 +189,6 @@ export default {
     // },
 
     isWelcomeBannerVisible () { return true },
-
-    upvoteElectionBanner () {
-      return {
-        title: this.I18n.t('pages.dho.home.communityElectionsAreAbout'),
-        description: this.I18n.t('pages.dho.home.weUseAFairAndInclusive')
-      }
-    },
 
     // TODO: waiting API
     // upvoteElectionBanner () {
@@ -251,117 +235,16 @@ export default {
   },
 
   created () {
-    if (localStorage.getItem('showUpvoteBanner') === 'false') {
-      this.isUpVoteElectionBannerVisible = false
-    }
-    this.counterdown = setInterval(() => {
-      this.formatTimeLeft()
-      this.$forceUpdate()
-    }, 1000)
     // TODO: waiting API
     // this.$apollo.queries.upvoteElection.refetch()
   },
 
-  beforeDestroy () {
-    clearInterval(this.counterdown)
-  },
-
-  methods: {
-    formatTimeLeft () {
-      const MS_PER_DAY = 1000 * 60 * 60 * 24
-      const MS_PER_HOUR = 1000 * 60 * 60
-      const MS_PER_MIN = 1000 * 60
-      const MS = 1000
-      const timeRemaining = this.votingTimeLeft()
-      if (timeRemaining > 0) {
-        const days = Math.floor(timeRemaining / MS_PER_DAY)
-        let lesstime = timeRemaining - (days * MS_PER_DAY)
-        const hours = Math.floor(lesstime / MS_PER_HOUR)
-        lesstime = lesstime - (hours * MS_PER_HOUR)
-        const min = Math.floor(lesstime / MS_PER_MIN)
-        lesstime = lesstime - (min * MS_PER_MIN)
-        const sec = Math.floor(lesstime / MS)
-        this.timeRemaining = {
-          days: days,
-          hours: hours,
-          mins: min,
-          sec: sec
-        }
-        return {
-          days: days,
-          hours: hours,
-          mins: min,
-          sec: sec
-        }
-      }
-      return 0
-    },
-
-    votingTimeLeft () {
-      // TODO: waiting API
-
-      // if (!this.upvoteElection) return 0
-
-      // const end = this.upvoteElection?.upcomingElection?.length ? new Date(this.upvoteElection?.startTime) : new Date(this.upvoteElection?.endTime)
-      const end = new Date(this.endDate)
-      const now = Date.now()
-      const t = end - now
-      // if (t < 0) {
-      //   this.$apollo.queries.upvoteElection.refetch()
-      // }
-      return t
-    },
-
-    hideUpvoteBanner () {
-      localStorage.setItem('showUpvoteBanner', false)
-      this.isUpVoteElectionBannerVisible = false
-    }
-  }
+  methods: {}
 }
 </script>
 
 <template lang="pug">
 q-page.page-dashboard
-  base-banner.q-mb-md(v-bind="upvoteElectionBanner" @onClose="hideUpvoteBanner" upvoteBanner :background="require('~/assets/images/election-banner-bg.jpeg')" v-if="isUpVoteElectionBannerVisible")
-    template(v-slot:right)
-      .flex.full-width.full-height.items-center.justify-center
-        q-card.q-pa-xl(:style="{ 'width': '350px', 'opacity': '.7', 'border-radius': '15px' }")
-          .col
-            .row.full-width.justify-center.q-pb-md(:style="{ 'border-bottom': '1px solid #242f5d'}")
-              .row
-                .time.row
-                  .row.items-center(v-if="timeRemaining.days > 0")
-                    .h-h4 {{ timeRemaining.days }}
-                    .h-h4.q-mx-xxs(v-if="timeRemaining.days > 1") {{ $t('pages.dho.home.days') }} :
-                    .h-h4.q-mx-xxs(v-else) {{ $t('pages.dho.home.day') }} :
-                  .row.items-center
-                    .h-h4 {{ timeRemaining.hours }}
-                    .h-h4.q-mx-xxs(v-if="timeRemaining.hours > 1") {{ $t('pages.dho.home.hours') }} :
-                    .h-h4.q-mx-xxs(v-else) {{ $t('pages.dho.home.hour') }} :
-                  .row.items-center
-                    .h-h4 {{ timeRemaining.mins }}
-                    .h-h4.q-mx-xxs(v-if="timeRemaining.mins > 1") {{ $t('pages.dho.home.mins') }}
-                    .h-h4.q-mx-xxs(v-else) {{ $t('pages.dho.home.min') }}
-                    .h-h4.q-mr-xxs(v-if="!timeRemaining.days > 0") :
-                  .row.items-center(v-if="!timeRemaining.days > 0")
-                    .h-h4 {{ timeRemaining.sec }}
-                    .h-h4.q-mx-xxs(v-if="timeRemaining.sec > 1") {{ $t('pages.dho.home.sec') }}
-                    .h-h4.q-mx-xxs(v-else) {{ $t('pages.dho.home.sec') }}
-              .row
-                .text-secondary {{ $t('pages.dho.home.timeLeftForSigningUp') }}
-            .row.q-pt-md
-              .col.flex.justify-center(:style="{ 'border-right': '1px solid #242f5d' }")
-                .h-h6 {{ participants }}
-                .full-width.flex.justify-center.text-secondary {{ $t('pages.dho.home.participants') }}
-              .col.flex.justify-center
-                .h-h6 $ {{ formatNumber(treasury) }}
-                .full-width.flex.justify-center.text-secondary {{ $t('pages.dho.home.treasury') }}
-    template(v-slot:buttons)
-      .row.justify-start
-        .flex(:class=" { 'q-mt-md': $q.screen.lt.md, 'justify-end': $q.screen.gt.sm }")
-        router-link(:to="{ name: 'upvote-election' }")
-          q-btn.q-px-lg.h-btn1(:class="{ 'q-mt-sm': $q.screen.lt.xs || $q.screen.xs }" no-caps rounded unelevated :label="$t('pages.dho.home.signup')" color="secondary" text-color="white")
-          q-btn.q-px-lg.h-btn1.q-ml-sm(color="white" flat :label="$t('pages.dho.home.learnMore')" no-caps rounded)
   base-banner(:compact="!$q.screen.gt.sm" :split="$q.screen.gt.md" v-bind="welcomeBanner" v-if="isWelcomeBannerVisible")
     template(v-slot:buttons)
       router-link(:to="{ name: 'organization' }")
